templates: Add support for `go-import` meta tag
These changes introduce the ability to add the necessary HTML meta tag
for Golang vanity import paths.

  <meta name="go-import" content="import-prefix vcs repo-root">

It can be configured by including the following frontmatter in every
Markdown file in the `/content/go-import` directory.

  [extra]
  vcs = ""
  repo_root = ""

The `import-prefix` specifier describes the import path that
corresponds to the root directory of the repository. It is
automatically derived from the Zola variable `current_url`.

The `vcs` identifier describes the version control system used. Allowed
values are `bzr`, `fossil`, `git`, `hg`, `svn`.

The `repo_root` identifier describes the root of the version control
system, which contains a schema and not a `.vcs` identifier.

Link: https://pkg.go.dev/cmd/go#hdr-Remote_import_paths
Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>