netlify: Migrate from Netlify back to webgo
On 27 February 2024, a call for help was published on HackerNews [1].
The reason for this was that, following a DDOS attack, a user received
a bill totalling $104,500.00 from Netlify. A total of 190TB of
bandwidth was used within 4 days, with Netlify charging $55/100GB. The
user claimed to have hosted an audio file (mp3) with a size of 3.44MB
on his website. The DDOS attack caused this file to be downloaded
multiple times.

Officially, Netlify currently has no DDOS protection [2]. It is also
not possible to limit the resources that the hosted project can
consume [3]. It is therefore possible that costs of a similar amount may be
incurred again at any time.

Currently, a bandwidth of approx. 4GB is used for the domain
"stefanco.de". The free maximum is 100GB. This may be due to the
embedded image files in the Jupyter notebooks (static/jupyter/...) and
the hosting of the BookPrint CLI (content/go-import/...).

To minimise the risk of unexpected costs, this project will be migrated
from Netlify back to webgo. The conditions of webgo include unlimited
bandwidth.

Link: https://news.ycombinator.com/item?id=39520776 [1]
Link: https://answers.netlify.com/t/limit-bandwidth-to-avoid-high-billing-caused-by-ddos/13086/3 [2]
Link: https://answers.netlify.com/t/limit-bandwidth-to-avoid-high-billing-caused-by-ddos/13086/5 [3]
Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>